feat: ChargeObject Decimal -> Int64

+++ CrypMulator/CrypMulator/Model/DTO/Realm/ChargeDTO.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 struct ChargeDTO {
-    let amount: Decimal
+    let amount: Int64
     let timestamp: Int64
 }
